requirements.txt
# Core ML dependencies
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=1.10.0
torchvision>=0.12.0
pytorch-lightning>=1.6.0
timm>=0.6.0
einops>=0.4.0
pytorch-metric-learning>=1.0.0

# Computer Vision
opencv-python>=4.8.1
Pillow>=10.1.0
albumentations>=1.3.0

# ML Ops & Utilities
wandb>=0.16.2
tensorboard>=2.15.0
tensorboardX>=2.6.0
hydra-core>=1.3.2
omegaconf>=2.3.0
tqdm>=4.66.1
pandas>=2.1.4
